% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGEDI.simulator-package.R
\docType{package}
\name{rGEDI.simulator-package}
\alias{rGEDI.simulator-package}
\title{rGEDI.simulator: A NASA's Global Ecosystem Dynamics Investigation (GEDI) simulator using ALS dataset.}
\description{
The rGEDI package provides functions for i) downloading, ii) visualizing, iii) clipping, iv) gridding, iv) simulating and v) exporting GEDI data.
}
\note{
See more details about GEDI data in \url{https://gedi.umd.edu/data/products/}.
}
\seealso{
For comprehensive examples refer to \url{https://github.com/carlos-alberto-silva/rGEDI/blob/master/README.md}
}
\author{
Carlos A. Silva, Caio Hamamura, Ruben Valbuena, Steve Hancock, Adrian Cardil, Eben N. Broadbent, Danilo R. A. de Almeida, Celso H. L. Silva Junior and Carine Klauberg
}
